refactored so all media types use the same tile declaration
