--- packages/shared-components/src/room/timeline/event-tile/MediaPreviewGroupView/AudioPreviewTile/AudioPreviewTile.module.css
@@ -1,13 +1,0 @@
-.bottom {
-    display: flex;
-    margin: var(--cpd-space-3x);
-}
-
-.tile {
-    border: 1px solid var(--cpd-color-border-interactive-secondary);
-    border-radius: 14px;
-
-    max-width: 100%;
-    overflow: clip;
-    background: var(--cpd-color-bg-subtle-secondary);
-}
