--- packages/shared-components/src/room/timeline/event-tile/MediaPreviewGroupView/ImagePreviewTile/ImagePreviewTile.tsx
@@ -1,19 +1,0 @@
-import React, { JSX } from "react";
-import { MediaPreviewGroupImageEntry } from "../MediaPreviewGroupView";
-import styles from "./ImagePreviewTile.module.css";
-import { Buttons, Icon, Image, LeftGroup, TextContent } from "../MediaPreviewComponents/MediaPreviewComponents";
-
-export function ImagePreviewTile(props: MediaPreviewGroupImageEntry): JSX.Element {
-    return (
-        <div className={styles.tile}>
-            <Image {...props} />
-            <div className={styles.bottom}>
-                <LeftGroup>
-                    <Icon {...props} />
-                    <TextContent {...props} />
-                </LeftGroup>
-                {props.buttons && props.buttons.length !== 0 && <Buttons buttons={props.buttons} />}
-            </div>
-        </div>
-    );
-}
